requirements.txt
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.12
pylast==7.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
six==1.17.0
spotipy==2.26.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.7.0
